Refactor imports and image paths for consistency
Refactored imports to consistently use the `common` module for shared resources. Updated image paths to use a dedicated `images` directory. This improves the structure and maintainability of the codebase.
