+++ common/shared.py
